Remove boilerplate for spinning up server in tests
