+++ packages/core/tests/integration/elicitations/elicitation-e2e.test.ts
@@ -3,33 +3,12 @@
 import { describe, expect, test } from "bun:test";
 import { z } from "zod";
 import { collectSseEventsCount } from "../../../../test-utils/src/sse.js";
-import {
-  InMemoryClientRequestAdapter,
-  InMemorySessionAdapter,
-  McpServer,
-  StreamableHttpTransport,
-} from "../../../src/index.js";
+import { createStatefulTestServer } from "../../utils.js";
 
 describe("Elicitation E2E Tests", () => {
   // FIXME - this does not actually test ctx.elicit
   test("E2E: ctx.elicit() does not throw when client has elicitation capability", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("with-elicitation", {
       description: "Test elicitation support",
@@ -103,23 +82,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: ctx.elicit() throws when client lacks elicitation capability", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("no-elicitation", {
       description: "Test no elicitation support",
@@ -206,23 +169,7 @@ describe("Elicitation E2E Tests", () => {
     // 4. Server resolves the elicitation promise with the result
     // 5. Tool completes with the elicitation data
 
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("full-elicitation", {
       description: "Test complete elicitation flow",
@@ -364,23 +311,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: full elicitation flow with client decline", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("decline-elicitation", {
       description: "Test elicitation decline flow",
@@ -515,23 +446,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: full elicitation flow with client cancel", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("cancel-elicitation", {
       description: "Test elicitation cancel flow",
@@ -663,23 +578,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: elicitation timeout when client doesn't respond", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("timeout-elicitation", {
       description: "Test elicitation timeout",
@@ -783,23 +682,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: elicitation with invalid client response data", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("invalid-response-elicitation", {
       description: "Test elicitation with invalid response",
@@ -938,23 +821,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: elicitation with client error response", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("error-response-elicitation", {
       description: "Test elicitation with client error",
@@ -1092,23 +959,7 @@ describe("Elicitation E2E Tests", () => {
     // The test expects both elicitation requests to be available simultaneously,
     // but sequential await calls mean the second elicitation doesn't start until
     // the first one completes. This requires more sophisticated event handling.
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("sequential-elicitations", {
       description: "Test multiple sequential elicitations",
@@ -1329,23 +1180,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: elicitation with empty schema", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("empty-schema-elicitation", {
       description: "Test elicitation with empty schema",
@@ -1478,23 +1313,7 @@ describe("Elicitation E2E Tests", () => {
   });
 
   test("E2E: elicitation with optional fields schema", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
+    const { server, handler } = createStatefulTestServer();
 
     server.tool("optional-fields-elicitation", {
       description: "Test elicitation with optional fields",
